Move encodeQueryName to its own file with dedicated tests
