+++ packages/plugins/apps/src/backend/discovery.ts
@@ -2,7 +2,6 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { createHash } from 'crypto';
 import type { Declaration, Identifier, Program } from 'estree';
 import type { AstNode } from 'rollup';
 
@@ -13,19 +12,6 @@ export interface BackendFunction {
     name: string;
     /** Absolute path to the .backend.ts source file */
     entryPath: string;
-}
-
-/**
- * Encode a BackendFunction into an opaque query name string.
- * Uses the full SHA-256 hash of the path so that backend file structure
- * is never leaked into frontend assets.
- *
- * This is the single source of truth for query name encoding — used by
- * proxy codegen, the production build, and the dev server.
- */
-export function encodeQueryName(ref: Pick<BackendFunction, 'path' | 'name'>): string {
-    const pathHash = createHash('sha256').update(ref.path).digest('hex');
-    return `${pathHash}.${ref.name}`;
 }
 
 /**
