+++ packages/plugins/apps/src/index.ts
@@ -10,7 +10,8 @@ import { createArchive } from './archive';
 import type { Asset } from './assets';
 import { collectAssets } from './assets';
 import type { BackendFunction } from './backend/discovery';
-import { encodeQueryName, extractExportedFunctions } from './backend/discovery';
+import { extractExportedFunctions } from './backend/discovery';
+import { encodeQueryName } from './backend/encodeQueryName';
 import { CONFIG_KEY, PLUGIN_NAME } from './constants';
 import { resolveIdentifier } from './identifier';
 import type { AppsOptions } from './types';
